This commit provides the capability of turning off use of the microphysics tendency above a user-specified height.  This change alleviates some instabilities encountered in deep-domain MPAS simulations (model tops above the stratopause).
The changes include:

(1) addition of a namelist variable specifying the height above which the
microphysics tendencies are to be ignored, and an integer variable that is set
when the integration starts specifying the level based on that height.
(2) additions to src/core_atmosphere/physics/mpas_atmphys_init.F to set the
level based on the namelist-specified height.
(3) changes in src/core_atmosphere/physics/mpas_atmphys_interface.F that disable
the microphysics tendancy updates above the specified level.